made changes for gcc-4.5 to compile with -m64 on a 64-bit system
git-svn-id: http://ctp2.darkdust.net/anonsvn/branches/linux@1113 34d2fc0c-33df-0310-8a34-85f8040ba873

+++ ctp2_code/gs/gameobj/Happy.cpp
@@ -996,7 +996,7 @@ double Happy::CalcFeats(Player *p)
 //
 //----------------------------------------------------------------------------
 void Happy::Copy(Happy *copy){
-	memcpy(&m_happiness, &copy->m_happiness, (uint32)&m_crime + sizeof(m_crime) - (uint32)&m_happiness);
+	memcpy(&m_happiness, &copy->m_happiness, (size_t)&m_crime + sizeof(m_crime) - (size_t)&m_happiness);
 	m_timedChanges = copy->m_timedChanges;
 
 	for (size_t i = 0; i < static_cast<size_t>(HAPPY_REASON_MAX); ++i){
